+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.1
+
+* Added `overrideStdout()`.
+* `await runZoned()`.
+* Renamed `run` to `overridePrint`, added a deprecated alias.
+
 ## 0.1.0
 
 * Initial release.